+++ example/lib/todo/todo_cube.dart
@@ -1,5 +1,4 @@
 import 'package:cubes/cubes.dart';
-import 'package:examplecube/todo/todo_register.dart';
 
 class TodoCube extends Cube with CubeNavigation {
   final todoList = ['first TODO', 'second TODO'].obs;
@@ -7,8 +6,4 @@ class TodoCube extends Cube with CubeNavigation {
   void addTodo(String todo) => todoList.add(todo);
 
   void delete(String todo) => todoList.remove(todo);
-
-  void navToRegister() {
-    navTo((context) => TodoRegister());
-  }
 }