+++ lib/src/actions/navigator_action.dart
@@ -18,15 +18,11 @@ enum NavigationType {
   pushNamed,
   pushNamedAndRemoveUntil,
   pushReplacementNamed,
-  push,
-  pushReplacement,
-  pushAndRemoveUntil,
   pop,
 }
 
-class NavigationAction extends CubeAction {
+class NavigationCubeAction extends CubeAction {
   final String? routeName;
-  final WidgetBuilder? builder;
   final Object? arguments;
   final RouteSettings? settings;
   final bool fullscreenDialog;
@@ -35,10 +31,9 @@ class NavigationAction extends CubeAction {
   final Object? result;
   final ValueChanged<Object?>? onResult;
 
-  NavigationAction({
-    this.type = NavigationType.push,
+  NavigationCubeAction({
+    this.type = NavigationType.pushNamed,
     this.routeName,
-    this.builder,
     this.arguments,
     this.onResult,
     this.settings,
@@ -47,40 +42,7 @@ class NavigationAction extends CubeAction {
     this.result,
   });
 
-  NavigationAction.push({
-    required this.builder,
-    this.settings,
-    this.onResult,
-    this.fullscreenDialog = false,
-  })  : type = NavigationType.push,
-        routeName = null,
-        arguments = null,
-        result = null,
-        predicate = null;
-
-  NavigationAction.pushReplacement({
-    required this.builder,
-    this.settings,
-    this.onResult,
-    this.fullscreenDialog = false,
-  })  : type = NavigationType.pushReplacement,
-        routeName = null,
-        arguments = null,
-        result = null,
-        predicate = null;
-
-  NavigationAction.pushAndRemoveUntil({
-    required this.builder,
-    required this.predicate,
-    this.settings,
-    this.onResult,
-    this.fullscreenDialog = false,
-  })  : type = NavigationType.pushAndRemoveUntil,
-        routeName = null,
-        result = null,
-        arguments = null;
-
-  NavigationAction.pop({
+  NavigationCubeAction.pop({
     this.result,
   })  : type = NavigationType.pop,
         routeName = null,
@@ -88,10 +50,9 @@ class NavigationAction extends CubeAction {
         settings = null,
         onResult = null,
         fullscreenDialog = false,
-        builder = null,
         arguments = null;
 
-  NavigationAction.pushNamed({
+  NavigationCubeAction.pushNamed({
     required this.routeName,
     this.arguments,
     this.onResult,
@@ -99,10 +60,9 @@ class NavigationAction extends CubeAction {
         predicate = null,
         settings = null,
         fullscreenDialog = false,
-        builder = null,
         result = null;
 
-  NavigationAction.pushNamedAndRemoveUntil({
+  NavigationCubeAction.pushNamedAndRemoveUntil({
     required this.routeName,
     required this.predicate,
     this.arguments,
@@ -110,10 +70,9 @@ class NavigationAction extends CubeAction {
   })  : type = NavigationType.pushNamedAndRemoveUntil,
         settings = null,
         fullscreenDialog = false,
-        builder = null,
         result = null;
 
-  NavigationAction.pushReplacementNamed({
+  NavigationCubeAction.pushReplacementNamed({
     required this.routeName,
     this.arguments,
     this.onResult,
@@ -121,7 +80,6 @@ class NavigationAction extends CubeAction {
         settings = null,
         predicate = null,
         fullscreenDialog = false,
-        builder = null,
         result = null;
 
   void handle(BuildContext context) {
@@ -143,34 +101,6 @@ class NavigationAction extends CubeAction {
       case NavigationType.pushReplacementNamed:
         context
             .goToNamedReplacement(routeName!, arguments: arguments)
-            .then((r) => onResult?.call(r));
-        break;
-      case NavigationType.push:
-        context
-            .goTo(
-              builder!,
-              settings: settings,
-              fullscreenDialog: fullscreenDialog,
-            )
-            .then((r) => onResult?.call(r));
-        break;
-      case NavigationType.pushReplacement:
-        context
-            .goToReplacement(
-              builder!,
-              settings: settings,
-              fullscreenDialog: fullscreenDialog,
-            )
-            .then((r) => onResult?.call(r));
-        break;
-      case NavigationType.pushAndRemoveUntil:
-        context
-            .goToAndRemoveUntil(
-              builder!,
-              predicate!,
-              settings: settings,
-              fullscreenDialog: fullscreenDialog,
-            )
             .then((r) => onResult?.call(r));
         break;
       case NavigationType.pop: