+++ lib/src/util/cube_navigation_mixin.dart
@@ -21,7 +21,7 @@ mixin CubeNavigation on Cube {
     ValueChanged<Object?>? onResult,
   }) {
     this.sendAction(
-      NavigationAction.pushNamed(
+      NavigationCubeAction.pushNamed(
         routeName: routeName,
         arguments: arguments,
         onResult: onResult,
@@ -36,7 +36,7 @@ mixin CubeNavigation on Cube {
     ValueChanged<Object?>? onResult,
   }) {
     this.sendAction(
-      NavigationAction.pushNamedAndRemoveUntil(
+      NavigationCubeAction.pushNamedAndRemoveUntil(
         routeName: routeName,
         arguments: arguments,
         predicate: predicate,
@@ -51,7 +51,7 @@ mixin CubeNavigation on Cube {
     ValueChanged<Object?>? onResult,
   }) {
     this.sendAction(
-      NavigationAction.pushReplacementNamed(
+      NavigationCubeAction.pushReplacementNamed(
         routeName: routeName,
         arguments: arguments,
         onResult: onResult,
@@ -59,59 +59,9 @@ mixin CubeNavigation on Cube {
     );
   }
 
-  void navTo(
-    WidgetBuilder builder, {
-    RouteSettings? settings,
-    bool fullscreenDialog = false,
-    ValueChanged<Object?>? onResult,
-  }) {
-    this.sendAction(
-      NavigationAction.push(
-        builder: builder,
-        fullscreenDialog: fullscreenDialog,
-        settings: settings,
-        onResult: onResult,
-      ),
-    );
-  }
-
-  void navToReplacement(
-    WidgetBuilder builder, {
-    RouteSettings? settings,
-    bool fullscreenDialog = false,
-    ValueChanged<Object?>? onResult,
-  }) {
-    this.sendAction(
-      NavigationAction.pushReplacement(
-        builder: builder,
-        fullscreenDialog: fullscreenDialog,
-        settings: settings,
-        onResult: onResult,
-      ),
-    );
-  }
-
-  void navToAndRemoveUntil(
-    WidgetBuilder builder,
-    RoutePredicate predicate, {
-    RouteSettings? settings,
-    bool fullscreenDialog = false,
-    ValueChanged<Object?>? onResult,
-  }) {
-    this.sendAction(
-      NavigationAction.pushAndRemoveUntil(
-        builder: builder,
-        fullscreenDialog: fullscreenDialog,
-        settings: settings,
-        predicate: predicate,
-        onResult: onResult,
-      ),
-    );
-  }
-
   void navPop<T>([T? result]) {
     this.sendAction(
-      NavigationAction.pop(result: result),
+      NavigationCubeAction.pop(result: result),
     );
   }
 }